requirements.txt
# Curi1 World Model - DreamerV4
# Python 依赖包

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 科学计算
numpy>=1.24.0
scipy>=1.10.0

# 强化学习环境
gym>=0.26.0
gymnasium>=0.28.0

# 图像处理
Pillow>=9.5.0
opencv-python>=4.7.0
imageio>=2.28.0

# 数据处理
h5py>=3.8.0
pandas>=2.0.0

# 配置管理
pyyaml>=6.0
omegaconf>=2.3.0

# 日志和可视化
tensorboard>=2.12.0
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.15.0  # 可选

# 进度条
tqdm>=4.65.0

# ROS 支持（可选，取决于实际机器人）
# rospy
# sensor-msgs
# std-msgs

# 其他工具
pytest>=7.3.0  # 测试
black>=23.3.0  # 代码格式化
flake8>=6.0.0  # 代码检查
